refactored tests and performed cs fix
