+++ app/DataDomain/Entities/Fiuselist/FiuselistEntity.php
@@ -33,5 +33,4 @@ class FiuselistEntity
     {
         $this->fiuselistEntries = $fiuselistEntries;
     }
-
 }
