+++ app/Exceptions/NotCreatedException.php
@@ -1,8 +1,6 @@
 <?php
 
-
 namespace App\Exceptions;
-
 
 use Exception;
 use Throwable;
