+++ app/Http/Controllers/Controller.php
@@ -11,7 +11,9 @@ use Illuminate\Support\Facades\Validator;
 
 class Controller extends BaseController
 {
-    use AuthorizesRequests, DispatchesJobs, ValidatesRequests;
+    use AuthorizesRequests;
+    use DispatchesJobs;
+    use ValidatesRequests;
 
     protected function validateRequest(RequestInterface $request): \Illuminate\Contracts\Validation\Validator
     {
