+++ app/Http/Requests/Content/ContentRequestInterface.php
@@ -1,10 +1,7 @@
 <?php
 
-
 namespace App\Http\Requests\Content;
-
 
 interface ContentRequestInterface
 {
-
 }
