+++ app/Http/Requests/Fiuselist/FiuselistAddContentRequest.php
@@ -1,15 +1,12 @@
 <?php
 
-
 namespace App\Http\Requests\Fiuselist;
-
 
 use App\Http\Requests\RequestInterface;
 use Illuminate\Foundation\Http\FormRequest;
 
 class FiuselistAddContentRequest extends FormRequest implements FiuselistRequestInterface, RequestInterface
 {
-
     public function rules(): array
     {
         return [
