+++ app/Http/Requests/Fiuselist/FiuselistRequestInterface.php
@@ -1,10 +1,7 @@
 <?php
 
-
 namespace App\Http\Requests\Fiuselist;
-
 
 interface FiuselistRequestInterface
 {
-
 }
