+++ app/Http/Requests/RequestInterface.php
@@ -1,8 +1,6 @@
 <?php
 
-
 namespace App\Http\Requests;
-
 
 interface RequestInterface
 {
