+++ app/Infrastructure/Repositories/Fiuselist/FiuselistRepositoryInterface.php
@@ -4,5 +4,4 @@ namespace App\Infrastructure\Repositories\Fiuselist;
 
 interface FiuselistRepositoryInterface
 {
-
 }
